Diego Ignacio Changes
I included a variable TIME which holds the time in miliseconds.
I also put mos of the main function code into a function which has the
possibility of running asyncronously with another function, although I
realized that such a thing might not be necessary. Hopefully it's
satisfying. If not, specifically tell me what is wrong and what is the
path I should take instead.